refactor(runtime/llm): extract completion-recovery helpers to own module
chat-executor-completion-recovery.ts created (415 LOC) covering:

  - attemptCompletionRecovery: the stop-hook blocking-message recovery
    path (push blocking msg, compact, re-call the model with optional
    tool_choice=required for mutation / artifact-evidence validation
    codes, surface exhausted / continuation outcomes).
  - attemptTokenBudgetContinuation: the continuation-controller token
    budget nudge path (inject nudge, compact, re-call).

Both return { recovered: boolean } instead of flipping an outer-scope
flag, so the caller owns shouldContinueAfterStopGate mutation
explicitly at the call sites.

Main loop drops 1,409 -> 1,070 LOC. Zero behavior delta - these were
closures capturing ctx / config / callbacks and inline helpers
resolveTurnOutputTokenBudget / shouldAllowBudgetContinuation which
are now inlined into the new module.

Tests green (9 pre-existing failures, zero new regressions).
